Header fixes
- remake logout modal
- changed styles of header elements
- transfer modals from old folder "components" to "modals"
