+++ src/components/Modals/UserLogoutModal/UserLogoutModal.module.css
@@ -1,0 +1,11 @@
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.80);
+  z-index: 100;
+}
+
+
